Added canvas system

+++ src/main/java/net/darktree/glslmc/render/PanoramaRenderer.java
@@ -2,7 +2,7 @@ package net.darktree.glslmc.render;
 
 public interface PanoramaRenderer {
 
-	void draw(float time, float mouseX, float mouseY, float width, float height, float alpha);
+	void draw(float time, float mouseX, float mouseY, int width, int height, float alpha);
 
 	void close();
